+++ lib/src/widgets/optional_string_field.dart
@@ -25,6 +25,9 @@ class OptionalStringField extends StatefulWidget {
   final TextInputType? keyboardType;
 
   final TextStyle? style;
+  final StrutStyle? strutStyle;
+  final TextAlign textAlign;
+  final TextAlignVertical? textAlignVertical;
   final TextInputAction? textInputAction;
   final TextFieldBuilder? builder;
   final bool? enabled;
@@ -47,6 +50,9 @@ class OptionalStringField extends StatefulWidget {
     this.onCleared,
     this.keyboardType,
     this.style,
+    this.strutStyle,
+    this.textAlign = TextAlign.start,
+    this.textAlignVertical,
     this.textInputAction,
     this.builder,
     this.enabled,
@@ -98,6 +104,9 @@ class _OptionalStringFieldState extends State<OptionalStringField> {
           controller: _controller,
           validator: widget.validator,
           style: widget.style,
+          strutStyle: widget.strutStyle,
+          textAlign: widget.textAlign,
+          textAlignVertical: widget.textAlignVertical,
           enabled: widget.enabled,
           readOnly: widget.readOnly,
           textInputAction: widget.textInputAction ??